Edit and Delete
Edit and Delete

+++ JS Web/JS Back-end/07. Exam preparation 2/Theater/controllers/playController.js
@@ -1,5 +1,6 @@
+const { default: mongoose } = require('mongoose');
 const { isUser } = require('../middlewares/guards');
-const { getAllPlays } = require('../services/play');
+const { getAllPlays, getPlayById } = require('../services/play');
 const { parseError } = require('../utils/parsers');
 
 const router = require('express').Router();
@@ -33,7 +34,7 @@ router.post('/create', isUser(), async (req,res) => {
                 public: Boolean(req.body.public)  
             }
         }
-        res.render('theater/create',ctx)
+        res.render('play/create',ctx)
     }
     
     
@@ -43,10 +44,12 @@ router.post('/create', isUser(), async (req,res) => {
 router.get('/details/:id', async (req,res) => {
     try {
     const play = await req.storage.getPlayById(req.params.id);
-    console.log(play)
+    console.log('Play',play.author)
     play.hasUser = Boolean(req.user);
+    
     play.isAuthor = req.user && req.user._id == play.author;
-    play.liked = req.user && play.userLiked.includes(req.user._id)
+    
+    play.liked = req.user && play.userLiked.find(u => u._id == req.user._id)
     
     res.render('play/details', {play})
     } catch(err) {
@@ -58,8 +61,13 @@ router.get('/details/:id', async (req,res) => {
     
 });
 
-router.get('/delete/:id', async (req,res) => {
+router.get('/delete/:id', isUser(), async (req,res) => {
     try {
+        const play = await req.storage.getPlayById(req.params.id);
+        
+        if(play.author != req.user._id) {
+            throw new Error('Cannot delete play you haven\'t created!')
+        }
         await req.storage.deletePlay(req.params.id);
         res.redirect('/');
     } catch(err) {
@@ -69,15 +77,59 @@ router.get('/delete/:id', async (req,res) => {
 });
 
 
-router.get('/edit:id', async (req,res) => {
-    // try {
-    //     await req.storage.deletePlay(req.params.id);
-    //     res.redirect('/');
-    // } catch(err) {
-    //     console.log(err.message);
-    //     res.redirect('play/details/' + req.params.id)
-    // }
+router.get('/edit/:id', isUser(), async (req,res) => {
+    try {
+        const play = await req.storage.getPlayById(req.params.id);
+        
+        
+        if(play.author != req.user._id) {
+            throw new Error('Cannot edit play you haven\'t created!')
+        }
+        
+        res.render('play/edit', {play})
+    } catch(err) {
+        console.log(err.message);
+        res.redirect('/play/details/' + req.params.id)
+    }
 });
+
+router.post('/edit/:id', isUser(), async (req,res) => {
+    try {
+        const play = await req.storage.getPlayById(req.params.id);
+        if(play.author != req.user._id) {
+            throw new Error('Cannot edit play you haven\'t created!')
+        }
+        await req.storage.editPlay(req.params.id, req.body);
+        res.redirect('/')
+    } catch(err) {
+        const ctx = {
+            errors: parseError(err),
+            play: {
+                _id: req.params.id,
+                title: req.body.title,
+                description: req.body.description,
+                imageUrl: req.body.imageUrl,
+                public: Boolean(req.body.public)
+            }
+        }
+        res.render('/play/edit',ctx)
+    };
+        
+     router.get('/like/:id',isUser(), async (req,res) => {
+        try {
+            const play = await req.storage.getPlayById(req.params.id);
+            
+            if(play.author == req.user._id) {
+                throw new Error('Cannot like you own play!')
+            }
+            await req.storage.likePlay(req.params.id,req.user._id);
+            res.redirect('/play/details/' + req.params.id);
+        } catch(err) {
+            console.log(err.message);
+            res.redirect('play/details/' + req.params.id)
+        }
+    });  
+})
